+++ NameplateAuraManager.lua
@@ -1,7 +1,9 @@
 local function defaults(classId)
 	local defaultConfigurations = {
-		[1] = {      -- Warrior
-			allowed = {
+		[1] = { -- Warrior
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {
 				[388539] = true, -- Rend
 				[275335] = true, -- Punish
 				[397364] = true, -- Thunderous Roar
@@ -17,14 +19,18 @@ local function defaults(classId)
 				[384954] = true, -- Shield Charge
 				[385042] = true -- Gushing Wound
 			},
-			blocked = {}
+			blockedDebuffs = {}
 		},
 		[2] = { -- Paladin
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
-		[3] = {      -- Hunter
-			allowed = {
+		[3] = { -- Hunter
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {
 				[217200] = true, -- Barbed Shot
 				[257284] = true, -- Hunter's Mark
 				[375893] = true, -- Death Chakram
@@ -35,47 +41,67 @@ local function defaults(classId)
 				[3355] = true, -- Freezing Trap
 				[195645] = true -- Wing Clip
 			},
-			blocked = {}
+			blockedDebuffs = {}
 		},
 		[4] = { -- Rogue
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[5] = { -- Priest
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[6] = { -- Death Knight
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[7] = { -- Shaman
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[8] = { -- Mage
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[9] = { -- Warlock
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[10] = { -- Monk
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[11] = { -- Druid
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[12] = { -- Demon Hunter
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		},
 		[13] = { -- Evoker
-			allowed = {},
-			blocked = {}
+			allowedBuffs = {},
+			blockedBuffs = {},
+			allowedDebuffs = {},
+			blockedDebuffs = {}
 		}
 	}
 
@@ -87,21 +113,33 @@ local function defaults(classId)
 end
 if not NAMDB then defaults() end
 
+local function shouldShowAura(aura, forceAll, allowedAuras, blockedAuras)
+	local playerAura = aura.sourceUnit == "player" or aura.sourceUnit == "pet" or aura.sourceUnit == "vehicle"
+	if next(allowedAuras) ~= nil or next(blockedAuras) ~= nil then
+		local returnValue = aura.nameplateShowAll or forceAll or
+			((allowedAuras[aura.spellId] and not blockedAuras[aura.spellId]) and playerAura)
+		return returnValue or false
+	else
+		return aura.nameplateShowAll or forceAll or (aura.nameplateShowPersonal and playerAura)
+	end
+end
+
 local function newShouldShowBuff(_, aura, forceAll)
-	if (not aura or not aura.name) then return false end
-    local _, _, classId = UnitClass("player")
+	if not aura or not aura.name then return false end
+	local _, _, classId = UnitClass("player")
 	local classDB = NAMDB[classId]
-	local hasCustomizations = next(classDB.allowed) ~= nil or next(classDB.blocked) ~= nil
-	return aura.nameplateShowAll or forceAll or
-		((hasCustomizations and (classDB.allowed[aura.spellId] and not classDB.blocked[aura.spellId]) or aura.nameplateShowPersonal) and
-			(aura.sourceUnit == "player" or aura.sourceUnit == "pet" or aura.sourceUnit == "vehicle"))
+	if aura.isHarmful then
+		return shouldShowAura(aura, forceAll, classDB.allowedDebuffs, classDB.blockedDebuffs)
+	else
+		return shouldShowAura(aura, forceAll, classDB.allowedBuffs, classDB.blockedBuffs)
+	end
 end
 
 local function Mixin(baseFrame)
 	baseFrame.UnitFrame.BuffFrame.ShouldShowBuff = newShouldShowBuff
 end
 
-local function handleSpellCommand(spellIdString, targetList, command, className)
+local function handleSpellCommand(spellIdString, targetList, command, className, auraType)
 	if not spellIdString or spellIdString == "" then
 		print("NAM: No spell ID provided.")
 		return
@@ -118,7 +156,7 @@ local function handleSpellCommand(spellIdString, targetList, command, className)
 	end
 	targetList[spellId] = not targetList[spellId]
 	local status = targetList[spellId] and "added to" or "removed from"
-	local text = status .. " " .. className .. " " .. command .. " list."
+	local text = status .. " " .. className .. " " .. auraType .. " " .. command .. " list."
 	print("NAM: " .. spellName .. " (" .. tostring(spellId) .. ") " .. text)
 end
 
@@ -126,32 +164,48 @@ SLASH_NAM1 = "/nam"
 SlashCmdList["NAM"] = function(msg)
 	local command, spellIdString = strsplit(" ", msg, 2)
 	command = string.lower(command)
-    local className, _, classId = UnitClass("player")
+	local className, _, classId = UnitClass("player")
 	local classDB = NAMDB[classId]
-	if command == "allow" then
-		handleSpellCommand(spellIdString, classDB.allowed, "allow", className)
-	elseif command == "block" then
-		handleSpellCommand(spellIdString, classDB.blocked, "block", className)
+	if command == "allowbuff" then
+		handleSpellCommand(spellIdString, classDB.allowedBuffs, "allow", className, "buff")
+	elseif command == "blockbuff" then
+		handleSpellCommand(spellIdString, classDB.blockedBuffs, "block", className, "buff")
+	elseif command == "allowdebuff" then
+		handleSpellCommand(spellIdString, classDB.allowedDebuffs, "allow", className, "debuff")
+	elseif command == "blockdebuff" then
+		handleSpellCommand(spellIdString, classDB.blockedDebuffs, "block", className, "debuff")
 	elseif command == "list" then
-		print("NAM: Allowed spells for " .. className .. ":")
-		for i, _ in pairs(classDB.allowed) do
+		print("NAM: Allowed buff auras for " .. className .. ":")
+		for i, _ in pairs(classDB.allowedBuffs) do
 			print("  " .. GetSpellInfo(i) .. " (" .. i .. ")")
 		end
-		print("NAM: Blocked spells for " .. className .. ":")
-		for i, _ in pairs(classDB.blocked) do
+		print("NAM: Blocked buff auras for " .. className .. ":")
+		for i, _ in pairs(classDB.blockedBuffs) do
+			print("  " .. GetSpellInfo(i) .. " (" .. i .. ")")
+		end
+		print("NAM: Allowed debuff auras for " .. className .. ":")
+		for i, _ in pairs(classDB.allowedDebuffs) do
+			print("  " .. GetSpellInfo(i) .. " (" .. i .. ")")
+		end
+		print("NAM: Blocked deuff auras for " .. className .. ":")
+		for i, _ in pairs(classDB.blockedDebuffs) do
 			print("  " .. GetSpellInfo(i) .. " (" .. i .. ")")
 		end
 	elseif command == "clear" then
-		classDB.allowed = {}
-		classDB.blocked = {}
+		classDB.allowedBuffs = {}
+		classDB.blockedBuffs = {}
+		classDB.allowedDebuffs = {}
+		classDB.blockedDebuffs = {}
 		print("NAM: Allow and block lists cleared for " .. className .. ".")
 	elseif command == "reset" then
 		defaults(classId)
 		print("NAM: Allow and block lists reset for " .. className .. ".")
 	elseif command == "help" or command == "?" or command == "" then
 		print("NAM: Commands:")
-		print("  `/nam allow [spellId]` to toggle an allowed aura.")
-		print("  `/nam block [spellid]` to toggle a blocked aura.")
+		print("  `/nam allowbuff [spellId]` to toggle an allowed aura on player nameplate.")
+		print("  `/nam blockbuff [spellid]` to toggle a blocked aura on player nameplate.")
+		print("  `/nam allowdebuff [spellId]` to toggle an allowed debuff on enemy nameplate.")
+		print("  `/nam blockdebuff [spellId]` to toggle a blocked debuff on enemy nameplate.")
 		print("  `/nam list` display class allow and block lists.")
 		print("  `/nam clear` clear class allow and block lists.")
 		print("  `/nam reset` reset class allow and block lists to default.")