+++ CHANGELOG.md
@@ -9,3 +9,6 @@
 - v2024.1
   - use game defaults (aura.nameplateShowPersonal) when empty or uninitialized class lists
   - added clear
+- v2024.2
+  - handle auras on player nameplate too
+  - sorry please delete your saved variables! this is my conversion function :)